requirements.txt
coverage==6.5.0
dvc==2.28.0
numpy==1.23.1
pandas==1.1.3
scikit-learn==1.1.1
pyyaml==5.3.1
Flask==2.0.1
#flask-cors==1.10.3
gunicorn==20.1.0
